feat(#584): moved all membership stuff to people
