+++ packages/epics/src/membership/components/index.ts
@@ -1,4 +1,3 @@
 export * from './outer-spaces';
 export * from './inner-spaces';
-export * from './members';
 export * from './subspace-detail';
