--- packages/epics/src/membership/components/members/index.ts
@@ -1,6 +1,0 @@
-export * from './member-card';
-export * from './members-list';
-export * from './members-section';
-export * from './member-head';
-export * from './member-spaces';
-export * from './member-detail';
