+++ packages/epics/src/people/components/index.ts
@@ -1,3 +1,9 @@
-export * from './button-profile';
 export * from './button-profile.connected';
+export * from './button-profile';
+export * from './member-card';
+export * from './member-detail';
+export * from './member-head';
+export * from './member-spaces';
+export * from './members-list';
+export * from './members-section';
 export * from './person-head';
