+++ packages/epics/src/people/components/member-card.tsx
