+++ packages/epics/src/people/components/member-head.tsx
