+++ packages/epics/src/people/components/members-list.stories.tsx
@@ -1,27 +1,19 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { MembersSection } from './members-section';
+import { MembersList } from './members-list';
 import { within } from '@storybook/testing-library';
 import { expect } from '@storybook/jest';
 
-const meta: Meta<typeof MembersSection> = {
-  component: MembersSection,
-  title: 'Epics/Membership/MembersSection',
+const meta: Meta<typeof MembersList> = {
+  component: MembersList,
+  title: 'Epics/People/MembersList',
 };
 export default meta;
-type Story = StoryObj<typeof MembersSection>;
+type Story = StoryObj<typeof MembersList>;
 
 export const Primary: Story = {
   args: {
     useMembers: () => ({
       isLoading: false,
-      pagination: {
-        hasNextPage: false,
-        hasPreviousPage: false,
-        page: 1,
-        pageSize: 10,
-        total: 2,
-        totalPages: 1,
-      },
       members: [
         {
           id: 1,
@@ -48,6 +40,9 @@ export const Primary: Story = {
   },
   play: async ({ canvasElement }) => {
     const canvas = within(canvasElement);
-    expect(canvas.getByText(/Members \|/gi)).toBeTruthy();
+    expect(canvas.getByText(/^Applicant/gi)).toBeTruthy();
+    expect(canvas.getByText(/^Active/gi)).toBeTruthy();
+    expect(canvas.getByText(/^Inactive/gi)).toBeTruthy();
+    expect(canvas.getByText(/^Rejected/gi)).toBeTruthy();
   },
 };
