+++ packages/epics/src/people/hooks/types.ts
@@ -1,4 +1,5 @@
 import { Person } from '@hypha-platform/core';
+import { FilterParams, PaginationMetadata } from '@hypha-platform/core';
 
 interface UseMeProps {
   newUserRedirectPath?: string;
@@ -9,3 +10,23 @@ export type UseMeReturn = {
   isLoading: boolean;
 };
 export type UseMe = (props: UseMeProps) => UseMeReturn;
+
+export type UseMembersReturn = {
+  members: Person[];
+  pagination?: PaginationMetadata;
+  isLoading: boolean;
+};
+
+export type UseMembersProps = {
+  page?: number;
+  filter?: FilterParams<Person>;
+};
+
+export type UseMembers = (props: UseMembersProps) => UseMembersReturn;
+
+export type UseMemberSpacesReturn = {
+  members: Person[];
+  pagination?: PaginationMetadata;
+  isLoading: boolean;
+};
+export type UseMemberSpaces = () => UseMemberSpacesReturn;
